Broke out FormHandler and Renamed CookieHandler
* Each Request method has its own class for FormHandler
* Renamed CookieHandler to specify which requests it can handle
