+++ src/main/java/http_server/FormHandlerPut.java
@@ -1,0 +1,22 @@
+package http_server;
+
+import java.io.IOException;
+
+public class FormHandlerPut implements Handler {
+  private DataStore dataStore;
+  private String parameter = "data";
+
+  public FormHandlerPut(DataStore dataStore) {
+    this.dataStore = dataStore;
+  }
+
+  public Response generate(Request request) throws IOException {
+    String parameterValue = request.getBodyParam(parameter);
+    dataStore.storeEntry(parameter, parameterValue);
+    Response response = new ResponseBuilder()
+                .setHttpVersion("HTTP/1.1")
+                .setStatusCode(200)
+                .build();
+    return response;
+  }
+}
